backend/requirements.txt
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
attrs==23.2.0
backoff==2.2.1
blis==0.7.11
boilerpy3==1.0.7
boto3==1.34.23
botocore==1.34.23
catalogue==2.0.10
cattrs==23.2.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
elastic-transport==7.16.0
elasticsearch==7.17.9
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
Events==0.5
faiss-cpu==1.7.4
farm-haystack==1.24.0
fastapi==0.109.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
h11==0.14.0
haystack-ai==2.0.0b5
haystack-bm25==1.0.2
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
inflect==7.0.0
install==1.3.5
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.2
langchain-community==0.0.16
langchain-core==0.1.16
langchain-openai==0.0.5
langcodes==3.3.0
langsmith==0.0.83
lazy-imports==0.3.1
llama-index==0.9.39
MarkupSafe==2.1.4
marshmallow==3.20.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.3
openai==1.10.0
opensearch-py==2.4.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.0
posthog==3.3.4
preshed==3.0.9
prompthub-py==4.0.0
protobuf==4.25.2
psutil==5.9.8
pyarrow==15.0.0
pydantic==1.10.14
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
rpds-py==0.17.1
s3transfer==0.10.0
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
sseclient-py==1.8.0
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
url-normalize==1.4.3
urllib3==1.26.18
uvicorn==0.26.0
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
